requirements.txt
python-telegram-bot==20.7
anthropic
openai
fastapi
uvicorn[standard]
aiohttp
fal-client
psycopg2-binary
python-dotenv
